refactor: 공통 내용 리팩토링 후 일부 파일 삭제
