--- src/types/message.ts
@@ -1,6 +1,0 @@
-export interface Message {
-  id: string;
-  text: string;
-  userId: string;
-  createdAt: Date;
-}
